+++ Generic/Generic/Program.cs
@@ -8,7 +8,7 @@ namespace Generic
             Console.WriteLine("Find maximum");
             int intOutput = MaxNumCheck.MaximunIntegerNumber(22, 66, 99);
             Console.WriteLine(intOutput);
-            double floatOutput = MaxNumCheck.MaximunFloatNumber(229.78, 112.96, 96.9);
+            float floatOutput = MaxNumCheck.MaximunFloatNumber(229.78f, 112.96f, 96.9f);
             Console.WriteLine(floatOutput);
             string stringOutput = MaxNumCheck.MaximunString("Mango", "Pineapple", "Banama");
             Console.WriteLine(stringOutput);
